Adiciona Long ao id

+++ src/main/java/br/com/crud/gweb/repository/UsuarioRepository.java
@@ -8,5 +8,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @date: 23/11/18
  */
 public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
-    Usuario findAById(Integer id);
+    Usuario findAById(Long id);
 }
